requirements.txt
streamlit
transformers
torch
requests-html
google-generativeai
